refactor(doctor): split nginx fronted-dir resolution into its own file
hardening_check_nginx_zones.go grew to 355 lines with the prior commit,
over the 300-line engineering-standard cap (TestFileSizeBudgetNotExceeded
in internal/repoqa). Moves resolveNginxFrontedDir and the new
directory-planning logic (planNginxAudit/nginxAuditPlan) into
hardening_check_nginx_fronted_dir.go — "which directory do we audit" is a
distinct concern from "does that directory's config carry the right rate
limits", and gives a future ResolveServedNginxDir generalization a home.
No behavior change; hardening_check_nginx_zones.go is now 299 lines.
